src/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
openai>=1.50.0
httpx>=0.23.0,<0.28
python-docx>=1.1.0
openpyxl>=3.1.5
aiosqlite>=0.20.0
python-multipart>=0.0.9
pydantic>=2.9.0
pydantic-settings>=2.5.0
fpdf2>=2.8.0
cryptography>=41.0.0
playwright>=1.40.0
